+++ tests/test_gff_parsers.py
@@ -1,0 +1,56 @@
+
+import os
+from inspect import getsourcefile
+import unittest
+from cdot.gff.gff_parser import GTFParser, GFF3Parser
+
+
+class Test(unittest.TestCase):
+    this_file_dir = os.path.dirname(os.path.abspath(getsourcefile(lambda: 0)))
+    test_data_dir = os.path.join(this_file_dir, "test_data")
+    ENSEMBL_GTF_FILENAME = os.path.join(test_data_dir, "ensembl_test.GRCh38.104.gtf")
+    REFSEQ_GFF3_FILENAME = os.path.join(test_data_dir, "refseq_test.GRCh38.p13_genomic.109.20210514.gff")
+    UCSC_GTF_FILENAME = os.path.join(test_data_dir, "hg19_chrY_300kb_genes.gtf")
+    FAKE_URL = "http://fake.url"
+
+    def _test_exon_length(self, transcripts, genome_build, transcript_id, expected_length):
+        transcript = transcripts[transcript_id]
+        exons = transcript["genome_builds"][genome_build]["exons"]
+        length = sum([exon[1] - exon[0] for exon in exons])
+        self.assertEquals(expected_length, length, "%s exons sum" % transcript_id)
+
+    def test_ucsc_gtf(self):
+        genome_build = "GRCh37"
+        parser = GTFParser(self.UCSC_GTF_FILENAME, genome_build, self.FAKE_URL)
+        _, transcripts = parser.get_genes_and_transcripts()
+        self._test_exon_length(transcripts, genome_build, "NM_013239", 2426)
+
+    def test_ensembl_gtf(self):
+        genome_build = "GRCh38"
+        parser = GTFParser(self.ENSEMBL_GTF_FILENAME, genome_build, self.FAKE_URL)
+        _, transcripts = parser.get_genes_and_transcripts()
+        self._test_exon_length(transcripts, genome_build, "ENST00000357654.9", 7088)
+
+    def test_refseq_gff3(self):
+        genome_build = "GRCh38"
+        parser = GFF3Parser(self.REFSEQ_GFF3_FILENAME, genome_build, self.FAKE_URL)
+        _, transcripts = parser.get_genes_and_transcripts()
+        self._test_exon_length(transcripts, genome_build, "NM_007294.4", 7088)
+
+    def test_exons_in_genomic_order(self):
+        genome_build = "GRCh38"
+        parser = GTFParser(self.ENSEMBL_GTF_FILENAME, genome_build, self.FAKE_URL)
+        _, transcripts = parser.get_genes_and_transcripts()
+        transcript = transcripts["ENST00000357654.9"]
+        exons = transcript["genome_builds"][genome_build]["exons"]
+        first_exon = exons[0]
+        last_exon = exons[-1]
+        self.assertGreater(last_exon[0], first_exon[0])
+
+        parser = GFF3Parser(self.REFSEQ_GFF3_FILENAME, genome_build, self.FAKE_URL)
+        _, transcripts = parser.get_genes_and_transcripts()
+        transcript = transcripts["NM_007294.4"]
+        exons = transcript["genome_builds"][genome_build]["exons"]
+        first_exon = exons[0]
+        last_exon = exons[-1]
+        self.assertGreater(last_exon[0], first_exon[0])